% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{addHighlyExpressedEdges}
\alias{addHighlyExpressedEdges}
\title{Add reactions without highly changing genes but with high average expression}
\usage{
addHighlyExpressedEdges(m, g, top = 3000)
}
\arguments{
\item{m}{Metabolic module}

\item{g}{Scored graph}
}
\description{
Add reactions without highly changing genes but with high average expression
}
