% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{collapseAtomsIntoMetabolites}
\alias{collapseAtomsIntoMetabolites}
\title{Collapse atoms belonging to the same metabolite into one vertex}
\usage{
collapseAtomsIntoMetabolites(m)
}
\arguments{
\item{m}{Metabolic module}
}
\value{
module in which atoms of the same metabolite are collapsed into one
}
\description{
Collapse atoms belonging to the same metabolite into one vertex
}
\examples{
data(mEx)
m <- collapseAtomsIntoMetabolites(m = mEx)

}
