% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{getGeneDEMeta}
\alias{getGeneDEMeta}
\title{Finds columns in gene differential expression table
required for gatom analysis}
\usage{
getGeneDEMeta(gene.de.raw, org.gatom.anno, idColumn = NULL, idType = NULL,
  pvalColumn = NULL, logPvalColumn = NULL, log2FCColumn = NULL,
  baseMeanColumn = NULL, signalColumn = NULL, signalRankColumn = NULL)
}
\arguments{
\item{gene.de.raw}{A table with differential expression results, an object
convertable to data.frame.}

\item{org.gatom.anno}{Organsim-specific annotation obtained from
makeOrgGatomAnnotation function}

\item{baseMeanColumn}{could be NULL (automatic), NA (no such column),
character (coumn name)}

\item{signalColumn}{could be NULL (automatic), character (coumn name)
function (evaluated in a scope of original data frame)}
}
\description{
Finds columns in gene differential expression table
required for gatom analysis
}
\examples{
data("org.Mm.eg.gatom.annoEx")
data("gene.de.rawEx")
de.meta <- getGeneDEMeta(gene.de.rawEx, org.gatom.anno = org.Mm.eg.gatom.annoEx)
de <- prepareDE(gene.de.rawEx, de.meta)
}
