% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getMetabolicPathways}
\alias{getMetabolicPathways}
\title{Generate list of metabolic pathways from Reactome and KEGG databases}
\usage{
getMetabolicPathways(
  universe,
  keggOrgCode,
  includeReactome = TRUE,
  includeKEGG = TRUE
)
}
\arguments{
\item{universe}{list of gene}

\item{keggOrgCode}{KEGG organism code, like mmu or hsa}

\item{includeReactome}{whether to include Reactome pathways (only works for Entrez ID universe)}

\item{includeKEGG}{wheter to include KEGG pahtways and modules}
}
\description{
Generate list of metabolic pathways from Reactome and KEGG databases
}
