% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{getMetDEMeta}
\alias{getMetDEMeta}
\title{Finds columns in differential expression table for metabolites
required for gatom analysis}
\usage{
getMetDEMeta(
  met.de.raw,
  met.db,
  idColumn = NULL,
  idType = NULL,
  pvalColumn = NULL,
  logPvalColumn = NULL,
  log2FCColumn = NULL,
  signalColumn = NULL
)
}
\description{
Finds columns in differential expression table for metabolites
required for gatom analysis
}
