% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getPathways2annotation}
\alias{getPathways2annotation}
\title{Adds pathway list to organism annotation object}
\usage{
getPathways2annotation(org.gatom.anno, organism = c("mmu", "hsa"))
}
\arguments{
\item{org.gatom.anno}{Organism annotation object}

\item{organism}{Which organism annotation object refers to (might be either mouse "mmu", either human "hsa")}
}
\description{
Adds pathway list to organism annotation object
}
