% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{gEx}
\alias{gEx}
\title{Example metabolic graph with atom topology.}
\format{
igraph object
}
\description{
See file \url{https://github.com/ctlab/gatom/blob/master/inst/scripts/example.R}
for details.
}
