% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{makeOrgGatomAnnotation}
\alias{makeOrgGatomAnnotation}
\title{Create an organism annotation object for network analysis}
\usage{
makeOrgGatomAnnotation(
  org.db,
  idColumns = c(Entrez = "ENTREZID", RefSeq = "REFSEQ", Ensembl = "ENSEMBL", Symbol =
    "SYMBOL"),
  nameColumn = "SYMBOL",
  enzymeColumn = "ENZYME",
  appendEnzymesFromKegg = TRUE,
  keggOrgCode = NULL
)
}
\arguments{
\item{org.db}{Bioconductor org.db object, e.g. org.Mm.eg.db}

\item{idColumns}{vector of column names from `org.db` object to cread ID mappings.
First ID will be used as a base identifier, should be compatible
with KEGG and Reactome databses.}

\item{nameColumn}{column with a human readable gene symbol. Default to "SYMBOL".}

\item{enzymeColumn}{column with an Enzyme Commission ID. Defatult to "ENZYME".}

\item{appendEnzymesFromKegg}{if TRUE, KEGG databases will be sued to extend gene to
enzyme mappings obtained from org.db package.}

\item{keggOrgCode}{KEGG organizm code, e.g. "mmu". If set to NULL, the code is determined
automatically.}
}
\description{
Create an organism annotation object for network analysis
}
