% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{met.de.rawEx}
\alias{met.de.rawEx}
\title{Example metabolite differential abundance data.}
\format{
tibble/data.frame object
}
\description{
See file https://github.com/ctlab/gatom/blob/master/data-raw/example.R for details
}
