% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{met.kegg.dbEx}
\alias{met.kegg.dbEx}
\title{Example KEGG-based metabolite database object}
\format{
list object
}
\description{
See file \url{https://github.com/ctlab/gatom/blob/master/inst/scripts/example.R}
for details.
}
