% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{networkEx}
\alias{networkEx}
\title{Example KEGG-based network object}
\format{
list object
}
\description{
See file \url{https://github.com/ctlab/gatom/blob/master/inst/scripts/example.R}
for details.
}
