% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{org.Mm.eg.gatom.annoEx}
\alias{org.Mm.eg.gatom.annoEx}
\title{Example organism annotation object}
\format{
list object
}
\description{
See file \url{https://github.com/ctlab/gatom/blob/master/inst/scripts/example.R}
for details.
}
