% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{prepareDE}
\alias{prepareDE}
\title{Makes data.table with differential expression results containing
all columns required for gatom in the expected format}
\usage{
prepareDE(de.raw, de.meta)
}
\description{
Makes data.table with differential expression results containing
all columns required for gatom in the expected format
}
