% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{prepareGeneDE}
\alias{prepareGeneDE}
\title{Makes data.table with gene expression table containing
all columns required for gatom in the expected format}
\usage{
prepareGeneDE(gene.de.raw, gene.de.meta)
}
\description{
Makes data.table with gene expression table containing
all columns required for gatom in the expected format
}

