% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_module_constructor.R
\name{run_module_constructor}
\alias{run_module_constructor}
\title{run_module_constructor}
\usage{
run_module_constructor(name, kpos_gene = 200, kpos_met = 200,
  r = 0.1, e = 0.1, ..., replace_with_pval_phosp = F, data = NA)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
run_module_constructor
}
