% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToDot}
\alias{saveModuleToDot}
\title{Save module to a graphviz dot file}
\usage{
saveModuleToDot(module, file, name = NULL)
}
\arguments{
\item{module}{Module to save}

\item{file}{File to save to}

\item{name}{Name of the module}
}
\description{
Save module to a graphviz dot file
}

