% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToPdf}
\alias{saveModuleToPdf}
\title{Save module to a nice pdf file}
\usage{
saveModuleToPdf(module, file, name = NULL, n_iter = 100, force = 1e-05)
}
\arguments{
\item{module}{Module to save}

\item{file}{File to save to}

\item{name}{Name of the module}

\item{n_iter}{Number of repel algorithm iterations}

\item{force}{Value of repel force}
}
\value{
Returns NULL
}
\description{
Save module to a nice pdf file
}
\examples{
data(mEx)
saveModuleToPdf(module = mEx, file = "module.pdf")

}
