% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToXgmml}
\alias{saveModuleToXgmml}
\title{Save module to an XGMML file}
\usage{
saveModuleToXgmml(module, file, name = NULL)
}
\arguments{
\item{module}{Module to save}

\item{file}{File to save to}

\item{name}{Name of the module}
}
\value{
Returns NULL
}
\description{
Save module to an XGMML file
}
\examples{
data(mEx)
saveModuleToXgmml(module = mEx, file = "module.xgmml")

}
