% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{styleWidget}
\alias{styleWidget}
\title{code adopted from https://github.com/ramnathv/htmlwidgets/issues/231}
\usage{
styleWidget(hw, style = "", addl_selector = "", elementId = NULL)
}
\description{
code adopted from https://github.com/ramnathv/htmlwidgets/issues/231
}
