% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{abbreviateLabels}
\alias{abbreviateLabels}
\title{Prepare lipid labels for lipid module}
\usage{
abbreviateLabels(module, orig.names, abbrev.names)
}
\arguments{
\item{module}{Module to prepare}

\item{orig.names}{whether to use original names from the dataset}

\item{abbrev.names}{whether to use abbreviated names for all lipids}
}
\description{
Prepare lipid labels for lipid module
}
