% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{addHighlyExpressedEdges}
\alias{addHighlyExpressedEdges}
\title{Add reactions without highly changing genes but with high average expression}
\usage{
addHighlyExpressedEdges(m, g, top = 3000)
}
\arguments{
\item{m}{Metabolic module}

\item{g}{Scored graph}

\item{top}{Maximum rank value for the gene to be considered highly expressed}
}
\value{
module with added edges that correspond to high average expression
}
\description{
Add reactions without highly changing genes but with high average expression
}
\examples{
data(mEx)
data(gEx)
m <- addHighlyExpressedEdges(m = mEx, g = gEx)

}
