% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{connectAtomsInsideMetabolite}
\alias{connectAtomsInsideMetabolite}
\title{Connect atoms belonging to the same metabolite with edges}
\usage{
connectAtomsInsideMetabolite(m)
}
\arguments{
\item{m}{Metabolic module}
}
\value{
module in which atoms of the same metabolite are connected
}
\description{
Connect atoms belonging to the same metabolite with edges
}
\examples{
data(mEx)
m <- connectAtomsInsideMetabolite(m = mEx)

}
