% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{createShinyCyJSWidget}
\alias{createShinyCyJSWidget}
\title{Creates shinyCyJS widget from module}
\usage{
createShinyCyJSWidget(
  module,
  layout = list(name = "cose-bilkent", animate = FALSE, randomize = FALSE,
    nodeDimensionsIncludeLabels = TRUE),
  ...
)
}
\description{
Creates shinyCyJS widget from module
}
