% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom-package.R
\docType{data}
\name{gene.de.rawEx}
\alias{gene.de.rawEx}
\title{Example gene differential expression data.}
\format{
tibble/data.frame object
}
\description{
See file \url{https://github.com/ctlab/gatom/blob/master/inst/scripts/example.R}
for details.
}
