% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{makeMetabolicGraph}
\alias{makeMetabolicGraph}
\title{Creates atom graph based on specified data}
\usage{
makeMetabolicGraph(
  network,
  topology = c("atoms", "metabolites"),
  org.gatom.anno,
  gene.de,
  gene.de.meta = getGeneDEMeta(gene.de, org.gatom.anno),
  gene.keep.top = 12000,
  met.db,
  met.de,
  met.de.meta = getMetDEMeta(met.de, met.db),
  met.to.filter = fread(system.file("mets2mask.lst", package = "gatom"))$ID,
  largest.component = TRUE
)
}
\arguments{
\item{network}{Network object}

\item{topology}{Way to determine network vertices}

\item{org.gatom.anno}{Organism annotation object}

\item{gene.de}{Table with the differential gene expression, set to NULL if absent}

\item{gene.de.meta}{Annotation of `gene.de` table}

\item{gene.keep.top}{Only the `gene.keep.top` of the most expressed genes will be kept for the network}

\item{met.db}{Metabolite database}

\item{met.de}{Table with the differential expression for metabolites, set to NULL if absent}

\item{met.de.meta}{Annotation of `met.de` table}

\item{met.to.filter}{List of metabolites to filter from the network}

\item{largest.component}{If TRUE, only the largest connected component is returned}
}
\description{
Creates atom graph based on specified data
}
