% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToDot}
\alias{saveModuleToDot}
\title{Save module to a graphviz dot file}
\usage{
saveModuleToDot(
  module,
  file,
  name = NULL,
  extra.node.attrs = NULL,
  extra.edge.attrs = NULL
)
}
\arguments{
\item{module}{Module to save}

\item{file}{File to save to}

\item{name}{Name of the module}

\item{extra.node.attrs}{Table with additional node attributes to be written to the dot file as is}

\item{extra.edge.attrs}{Table with additional edge attributes to be written to the dot file as is}
}
\value{
Returns NULL
}
\description{
Save module to a graphviz dot file
}
\examples{
data(mEx)
saveModuleToDot(module = mEx, file = "module.dot")

}
