% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToHtml}
\alias{saveModuleToHtml}
\title{Save module to a html widget}
\usage{
saveModuleToHtml(
  module,
  file,
  name = "",
  sizingPolicy = htmlwidgets::sizingPolicy(defaultWidth = "100\%", defaultHeight =
    "90vh", padding = 10),
  ...
)
}
\arguments{
\item{module}{Module to save}

\item{file}{File to save to}

\item{name}{Name of the module}
}
\description{
Save module to a html widget
}
