% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveModuleToPdf}
\alias{saveModuleToPdf}
\title{Save module to a nice pdf file}
\usage{
saveModuleToPdf(module, layout = "sna_kk", n_iter = 100, force = 1e-05,
  file, name = NULL)
}
\arguments{
\item{module}{Module to save}

\item{layout}{Edge layout}

\item{n_iter}{Number of repel algorithm iterations}

\item{force}{Value of repel force}

\item{file}{File to save to}

\item{name}{Name of the module}
}
\description{
Save module to a nice pdf file
}
