% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{scoreGraph}
\alias{scoreGraph}
\title{Score metabolic graph}
\usage{
scoreGraph(
  g,
  k.gene,
  k.met,
  vertex.threshold.min = 0.1,
  edge.threshold.min = 0.1,
  met.score.coef = 1,
  show.warnings = TRUE,
  raw = FALSE
)
}
\arguments{
\item{g}{Metabolic graph obtained with makeMetabolic graph function}

\item{k.gene}{Number of gene signals to be scored positively, the higher is the number,
the larger will be the resulting module. If set to NULL, genes will not be used for scoring.}

\item{k.met}{Number of metabolite signals to be scored positively, the higher is the number,
the larger will be the resulting module. If set to NULL, metabolites will not be used for scoring.}
}
\description{
Score metabolic graph
}
